Implemented up to Day 21
The inputAOC2018_21.txt is the *modified* version of my input. I've added a /sub/line to adjust the calculation of R3 to avoid the loop to find the number which is a division of (in my case) R4 and 256. To calculate the value I had to introduce a new opcode *divi* which does the integer division of a Register and an Immediate value.